+++ RELEASE_NOTES.md
@@ -1,11 +1,14 @@
-
+Version 0.2.2 (January 10th 2017)
+------------------------------
+ * added toggle() method to toggle between play and pause states
+ * play(), pause() and stop() methods now return a boolean
 
 Version 0.2.1 (October 15th 2016)
 ------------------------------
  * added performance.now() polyfill
  * uses performance.now instead of Date.getTime() for better performances
  * added Time class
- * added getRoundedDelta() function
+ * added getRoundedDelta() method
 
 Version 0.2.0 (September 21st 2016)
 ------------------------------